requirements.txt
google-cloud-aiplatform[adk,agent_engines]

python-dotenv

pydantic

PyYAML

google.genai